+++ experimental/reporting-ui/src/main/react/reporting-ui/BUILD.bazel
@@ -5,6 +5,22 @@ load("@aspect_rules_webpack//webpack:defs.bzl", "webpack_bundle", "webpack_devse
 npm_link_all_packages()
 
 js_library(
+    name = "src",
+    srcs = [
+        "app.css",
+        "app.tsx",
+        "index.css",
+        "index.tsx",
+        "logo.tsx",
+        "report_web_vitals.ts",
+    ],
+    visibility = [
+        "//experimental/reporting-ui/src/main/react/reporting-ui:__subpackages__",
+        "//experimental/reporting-ui/src/test/react/reporting-ui:__subpackages__",
+    ],
+)
+
+js_library(
     name = "build_node_deps",
     srcs = [
         "//:node_modules/babel-loader",
@@ -37,11 +53,12 @@ js_library(
 webpack_bundle(
     name = "webpack_bundle",
     srcs = [
+        ":src",
         ":webpack_config",
         "//:tsconfig",
-        "//experimental/reporting-ui/src/main/react/reporting-ui/library",
-        "//experimental/reporting-ui/src/main/react/reporting-ui/public",
-        "//experimental/reporting-ui/src/main/react/reporting-ui/src",
+        "//experimental/reporting-ui/src/main/react/reporting-ui/client",
+        "//experimental/reporting-ui/src/main/react/reporting-ui/model/reporting",
+        "//experimental/reporting-ui/src/main/react/reporting-ui/public"
     ],
     node_modules = "//:node_modules",
     output_dir = True,
@@ -55,10 +72,11 @@ webpack_devserver(
     name = "webpack_devserver",
     data = [
         ":build_node_deps",
+        ":src",
         "//:tsconfig",
-        "//experimental/reporting-ui/src/main/react/reporting-ui/library",
-        "//experimental/reporting-ui/src/main/react/reporting-ui/public",
-        "//experimental/reporting-ui/src/main/react/reporting-ui/src",
+        "//experimental/reporting-ui/src/main/react/reporting-ui/client",
+        "//experimental/reporting-ui/src/main/react/reporting-ui/model/reporting",
+        "//experimental/reporting-ui/src/main/react/reporting-ui/public"
     ],
     node_modules = "//:node_modules",
     webpack_config = ":webpack_config",
