--- experimental/reporting-ui/src/main/react/reporting-ui/library/reporting_client/BUILD.bazel
@@ -1,20 +1,0 @@
-load("@aspect_rules_js//js:defs.bzl", "js_library")
-
-package(
-    default_visibility = [
-        "//experimental/reporting-ui/src/main/react/reporting-ui/library:__subpackages__",
-    ],
-)
-
-js_library(
-    name = "reporting_client",
-    srcs = [
-        "memoize.ts",
-        "real_api.ts",
-        "reporting_client_impl.ts",
-        "reporting_client.ts",
-    ],
-    deps = [
-        "//experimental/reporting-ui/src/main/react/reporting-ui/library/reporting_client/models",
-    ],
-)
