+++ experimental/reporting-ui/src/main/react/reporting-ui/model/reporting/BUILD.bazel
@@ -1,14 +1,14 @@
 load("@aspect_rules_js//js:defs.bzl", "js_library")
-load("@aspect_rules_ts//ts:defs.bzl", "ts_project")
 
 package(
     default_visibility = [
-        "//experimental/reporting-ui/src/main/react/reporting-ui/library:__subpackages__",
+        "//experimental/reporting-ui/src/main/react/reporting-ui:__subpackages__",
+        "//experimental/reporting-ui/src/test/react/reporting-ui:__subpackages__",
     ],
 )
 
 js_library(
-    name = "models",
+    name = "reporting",
     srcs = [
         "demographic.ts",
         "get_report_request.ts",
