+++ experimental/reporting-ui/src/main/react/reporting-ui/model/reporting/summary_publisher_data.ts
