+++ experimental/reporting-ui/src/main/react/reporting-ui/public/assets/fonts/open-sans-light.ttf
