+++ experimental/reporting-ui/src/main/react/reporting-ui/public/assets/icons/options.tsx
