--- experimental/reporting-ui/src/main/react/reporting-ui/src/BUILD.bazel
@@ -1,20 +1,0 @@
-load("@aspect_rules_js//js:defs.bzl", "js_library")
-
-package(
-    default_visibility = [
-        "//experimental/reporting-ui/src/main/react/reporting-ui:__subpackages__",
-        "//experimental/reporting-ui/src/test/react/reporting-ui:__subpackages__",
-    ],
-)
-
-js_library(
-    name = "src",
-    srcs = [
-        "app.css",
-        "app.tsx",
-        "index.css",
-        "index.tsx",
-        "logo.tsx",
-        "report_web_vitals.ts",
-    ],
-)
