+++ experimental/reporting-ui/src/main/react/reporting-ui/webpack.config.js
@@ -16,7 +16,7 @@ const path = require('path');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 
 module.exports = {
-  entry: path.resolve(__dirname, './src/index.tsx'),
+  entry: path.resolve(__dirname, './index.tsx'),
   module: {
     rules: [
       {
