+++ experimental/reporting-ui/src/test/react/reporting-ui/fake/BUILD.bazel
@@ -5,9 +5,9 @@ webpack_devserver(
     data = [
         "//:tsconfig",
         "//experimental/reporting-ui/src/main/react/reporting-ui:build_node_deps",
-        "//experimental/reporting-ui/src/main/react/reporting-ui/library",
+        "//experimental/reporting-ui/src/main/react/reporting-ui:src",
+        "//experimental/reporting-ui/src/main/react/reporting-ui/model/reporting",
         "//experimental/reporting-ui/src/main/react/reporting-ui/public",
-        "//experimental/reporting-ui/src/main/react/reporting-ui/src",
         "//experimental/reporting-ui/src/test/react/reporting-ui/fake/src",
     ],
     node_modules = "//:node_modules",
